requirements.txt
pyyaml>=6.0
pandas>=2.0.0
openpyxl>=3.1.0
tqdm>=4.65.0
tenacity>=8.2.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
httpx>=0.27.0
playwright>=1.40.0
google-search-results>=2.4.2
